#15 Updateproduct

+++ shop/product/forms.py
@@ -1,10 +1,10 @@
-from wtforms import Form, SubmitField, IntegerField, FloatField, StringField, TextAreaField, validators
+from wtforms import Form, SubmitField, IntegerField, FloatField, StringField, TextAreaField, validators, DecimalField
 from flask_wtf.file import FileField, FileRequired, FileAllowed
 
 
 class Addproducts(Form):
     name = StringField('Name', [validators.DataRequired()])
-    price = FloatField('Price', [validators.DataRequired()])
+    price = DecimalField('Price', [validators.DataRequired()])
     discount = IntegerField('Discount', default=0)
     stock = IntegerField('Stock', [validators.DataRequired()])
     color = StringField('Color', [validators.DataRequired()])
